Trading

Trader

Trading.Trader — Type

Trader(broker::AbstractBroker; tickers::Vector{String}, strategies::Vector{Strategy}, start=current_time())

Holds all data and tasks related to trading.

Brokers

Trading.AbstractBroker — Type

AbstractBroker

Interface for external brokers.

Trading.AlpacaBroker — Type

AlpacaBroker

Broker to communicate with Alpaca. Can be constructed with your key_id and secret_key (see connect-to-alpaca-api).

Trading.HistoricalBroker — Type

HistoricalBroker

Stores and provides data from historical datasets. Data can be streamed fashion by assigning a Clock to the clock constructor kwarg, which will be used to determine the next bar to stream when calling receive on this broker.

Trading.MockBroker — Type

MockBroker

Mimics all function of a normal broker but with random data.

Portfolio

Data

Trading.Cash — Type

Cash

Represents the actual cash balance. Currently there is no particular currency tied to this.

Trading.PurchasePower — Type

PurchasePower

Represents the current purchasing power. This is updated at the start of each update cycle to the current value of the Cash singleton. It can thus be used to determine how many purchases/trades can be made during one cycle.

Trading.Position — Type

Position

Represents a position held in an equity represented by ticker.

Trading.Purchase — Type

Purchase

The local representation of a purchase order. This will be turned into an Order by the Purchaser System as soon as it's communicated to the broker.

Trading.Sale — Type

Purchase

The local representation of a sell order. This will be turned into an Order by the Seller System as soon as it's communicated to the broker.

Trading.Order — Type

Order

Representation of a Purchase or Sale order that has been communicated to the broker. Once the status goes to "filled" the filling information will be taken by the Filler System to create a Filled component.

Trading.Filled — Type

Filled

Represents the filled avg_price and quantity of an Order.

Trading.PortfolioSnapshot — Type

PortfolioSnapshot

A snapshot of the current Positions and Cash value of the portfolio.

Systems

Trading.Purchaser — Type

Purchaser

Handles Purchases. Mainly verifies prices and quantities to be purchased.

Trading.Seller — Type

Seller

Handles Sales.

Trading.Filler — Type

Filler

When the status of an Order changes to "filled", the filled quantity and average fill price is registered in a Filled Component.

Trading.DayCloser — Type

DayCloser([interval::Period = Minute(1)])

Closes the day. Will run during the time interval [market_close - interval, market_close]. Currently it just removes pending trades.

Trading.SnapShotter — Type

SnapShotter([interval::Period = Minute(1)])

Takes a PortfolioSnapshot after each interval, storing Positions, Cash, and total value.

Bars

Trading.Open — Type

Open

The opening price of a given bar.

Trading.High — Type

High

The highest price of a given bar.

Trading.Low — Type

Low

The lowest price of a given bar.

Trading.Close — Type

Close

The closing price of a given bar.

Trading.Volume — Type

Volume

The traded volume of a given bar.

Indicators

Data

Trading.SMA — Type

SMA{horizon, T}

The simple moving average of a value over a timeframe of horizon.

Trading.MovingStdDev — Type

MovingStdDev{horizon, T}

The moving standard deviation of a value over timeframe of horizon.

Trading.EMA — Type

EMA{horizon, T}

The exponential moving average of a value over timeframe of horizon.

Trading.RSI — Type

RSI{horizon, T}

The relative strength index of a value over timeframe of horizon.

Trading.Bollinger — Type

Bollinger{horizon, T}

The up and down Bollinger bands for a value, over a timeframe of horizon.

Trading.Sharpe — Type

Sharpe{horizon, T}

The sharpe ratio of a value over a timeframe horizon.

Systems

Trading.SMACalculator — Type

SMACalculator

Calculates the SMA of data.

Trading.MovingStdDevCalculator — Type

MovingStdDevCalculator

Calculates the MovingStdDev of data.

Trading.EMACalculator — Type

EMACalculator

Calculates the EMA of data.

Trading.RSICalculator — Type

RSICalculator

Calculates the RSI of values.

Trading.BollingerCalculator — Type

BollingerCalculator

Calculates the Bollinger bands for data. The width parameter can be tuned, by default it is 2.0.

Trading.SharpeCalculator — Type

SharpeCalculator

Calculates the bare Sharpe ratio of data.

General Data

Trading.LogVal — Type

LogVal

The logarithm of a value.

Trading.Difference — Type

Difference

The lag 1 difference.

Trading.RelativeDifference — Type

RelativeDifference

The lag 1 relative difference.

Trading.LogValCalculator — Type

LogValCalculator

Calculates the logarithms of data.

Trading.DifferenceCalculator — Type

DifferenceCalculator

Computes the lag 1 differences of data values.

Trading.RelativeDifferenceCalculator — Type

RelativeDifferenceCalculator

Calculates the lag 1 RelativeDifference of data.