Ticker Ledgers

Trading.AssetLedger — Type

AssetLedger

A AssetLedger holds the data for a given ticker as it arrives. Currently this is bar data in the form of Open, High, Low, Close and Volume, produced by a DataStream. If certain derived Indicator data is requested, it also holds this as it is produced by the different systems.

Trading.Asset — Type

Represents the various asset classes. So far Stock or Crypto, each with a ticker as identification.

Bar Components

Bars are represented internally by the following set of Components, basically standard ohlc and volume:

Open

Close

High

Low

Volume

Their differences (i.e. from one period to the next), and logarithms are represented by:

Difference

RelativeDifference

LogVal

Each of these, and some derived quantities, can be requested by a Strategy as discussed in Indicators and demonstrated in more detail in the Strategies tutorial.

Reference

Trading.Open — Type

Open

The opening price of a given bar.

Trading.High — Type

High

The highest price of a given bar.

Trading.Low — Type

Low

The lowest price of a given bar.

Trading.Close — Type

Close

The closing price of a given bar.

Trading.Volume — Type

Volume

The traded volume of a given bar.

Trading.Difference — Type

Difference

The lag 1 difference.

Trading.RelativeDifference — Type

RelativeDifference

The lag 1 relative difference.

Trading.LogVal — Type

LogVal

The logarithm of a value.

Trading.value — Function

value(b::SingleValIndicator)

Returns the number that is stored in the SingleValIndicator. This is by default the v field.