Data

Historical Acquisition

The current suggested method of historical data acquisition is through the AlpacaBroker. There are three types of data that can be retrieved: bars, quotes and trades.

Trading.bars — Function

bars(broker, ticker, start, stop; timeframe, kwargs...)

Retrieve the bar data for ticker from start to stop and with an interval of timeframe. When using AlpacaBroker see the Bar Object documentation for further reference.

Example

broker = AlpacaBroker(<key_id>, <secret_key>)

bars(broker, "AAPL", DateTime("2022-01-01T00:00:00"), DateTime("2023-01-01T00:00:00")l timeframe = Minute(1))

The above will retrieve 2022 bar data "AAPL" on a Minute resolution.

Trading.quotes — Function

quotes(broker, ticker, start, stop)

Returns the quotes made for ticker between start and stop. When using AlpacaBroker see the Quote Object documentation for further reference.

Example

broker = AlpacaBroker(<key_id>, <secret_key>)

quotes(broker, "AAPL", DateTime("2022-01-01T14:30:00"), DateTime("2022-01-01T14:31:00"))

Trading.trades — Function

trades(broker, ticker, start, stop)

Returns the trades made for ticker between start and stop. When using AlpacaBroker see the Trade Object documentation for further reference.

Example

broker = AlpacaBroker(<key_id>, <secret_key>)

trades(broker, "AAPL", DateTime("2022-01-01T14:30:00"), DateTime("2022-01-01T14:31:00"))

Each of these will return a TimeArray, which can be used e.g. with MarketTechnicals for specialized analysis. An AbstractBroker has an internal cache that will retain previously requested data.

Data Stream

There are realtime data streams for bars and portfolio/order updates. They follow the same semantics as the standard HTTP.WebSockets.WebSocket.

Trading.BarStream — Type

BarStream

Supplies a stream of bars from a broker. Can be created by calling bar_stream on an AbstractBroker. See receive and register! for more information.

Missing docstring.

Missing docstring for Trading.bar_stream. Check Documenter's build log for details.

Trading.OrderStream — Type

OrderStream

Interface to support executing trades and retrieving account updates.

Trading.order_stream — Function

order_stream(f::Function, broker::AbstractBroker)

Creates an OrderStream to stream order data. Uses the same semantics as a standard HTTP.WebSocket.

Example

broker = AlpacaBroker(<key_id>, <secret_key>)

order_stream(broker) do stream
    order = receive(stream)
end

Ticker Ledgers

Trading.TickerLedger — Type

TickerLedger

A TickerLedger holds the data for a given ticker as it arrives. Currently this is bar data in the form of Open, High, Low, Close and Volume, produced by a BarStream. If certain derived Indicator data is requested, it also holds this as it is produced by the different systems.