Tutorial

As with any other functionality in Trading, Strategies are represented by Systems and thus are treated on completely equal footing with the core functionality. There are three main parts that need to be implemented for a Strategy to be used: the Systems, the Overseer.update functions, and the Overseer.requested_components function. This latter one will be used to determine which Indicator systems need to be running on the TickerLedgers in order to produce the Indicators that are used by the Strategy. The update function of a Strategy System is ran periodically after the update functions for the other Systems that make the rest of the Trader tick.

Strategy Definition

As an example we will implement a very simple slow/fast moving average strategy, i.e. SlowFast. The goal is that we can later use it in our Trader in to following way:

trader = Trader(broker; strategies = [Strategy(:slowfast, [SlowFast()], tickers=["stock1", "stock2"])])

We begin by defining the SlowFast System and the components that it requests to be present in TickerLedgers. They will be automatically created as tick data arrives.

struct SlowFast <: System end

Overseer.requested_components(::SlowFast) = (SMA{50, Close}, SMA{200, Close})

We here request the slow and fast sma components of the closing price (SMA{200, Trading.Close}, SMA{50, Trading.Close}).

We then implement the following update function that will be executed periodically:

function Overseer.update(s::SlowFast, t::Trader, ticker_ledgers)
    for ticker_ledger in ticker_ledgers

        ticker = ticker_ledger.ticker

        for e in new_entities(ticker_ledger, s)
            lag_e = lag(e, 1)

            if lag_e === nothing
                continue
            end

            sma_50  = e[SMA{50, Close}].sma
            sma_200 = e[SMA{200, Close}].sma

            lag_sma_50 = lag_e[SMA{50, Close}].sma
            lag_sma_200 = lag_e[SMA{200, Close}].sma

            if sma_50 > sma_200 && lag_sma_50 < lag_sma_200
                Entity(t, Sale(ticker, 1.0))
            elseif sma_50 < sma_200 && lag_sma_50 > lag_sma_200
                Entity(t, Purchase(ticker, 1.0))
            end

        end
    end
end

Let's go through this line by line:

for ticker_ledger in ticker_ledgers

We loop through each of the ticker ledgers that this strategy was created for (i.e. stock1, stock2).

for e in new_entities(ticker_ledger, s)

Then, we ask for the new_entities in the TickerLedger that have in this case both the SMA{200, Close} and SMA{50,Close} components. Each of these entities will be touched once and only once.

lag_e = lag(e, 1)

Since we are looking for crossings between the two moving averages, we ask for the entity of the previous time. If there was none, i.e. e is the very first entity, lag(e, 1) will return nothing and so we don't do anything.

sma_50  = e[SMA{50, Close}].sma
sma_200 = e[SMA{200, Close}].sma

lag_sma_50 = lag_e[SMA{50, Close}].sma
lag_sma_200 = lag_e[SMA{200, Close}].sma

We retrieve the sma's for both the current and lagged entity.

if sma_50 > sma_200 && lag_sma_50 < lag_sma_200
    Entity(t, Sale(ticker, 1.0))
elseif sma_50 < sma_200 && lag_sma_50 > lag_sma_200
    Entity(t, Purchase(ticker, 1.0))
end

If the fast sma crosses above the slow sma, we assume the stock is overbought and we sell it by creating an Entity with a Sale component. Vice versa, If the fast sma crosses below the slow sma, we assume the stock is oversold and we buy it by creating an Entity with a Purchase component.

BackTesting

The framework is set up to treat backtesting and realtime trading in completely identical ways, and we can therefore backtest our strategy on some historical data.

We first define the broker from which to pull the historical data, in this case we use AlpacaBroker with our key_id and secret. We then use it in the HistoricalBroker which supplies data in the same way of a realtime broker would.

We then set up the strategy for the MSFT and AAPL tickers, define our BackTester with our data range and interval dt.

Note

When using daily data (e.g. dt=Day(1)), it is important to specify only_day=false, otherwise nothing will happen since our strategy will only run during trading hours, and no daily bars will have a timestamp inside those hours.

Finally we use start to loop through all the days and execute the strategy, possible trades, and any other behavior as if it is realtime.

broker = HistoricalBroker(AlpacaBroker(ENV["ALPACA_KEY_ID"], ENV["ALPACA_SECRET"]))

strategy = Strategy(:slowfast, [SlowFast()], tickers=["MSFT", "AAPL"])

trader = BackTester(broker, start = DateTime("2015-01-01T00:00:00"),
                            stop = DateTime("2020-01-01T00:00:00"),
                            dt = Day(1),
                            strategies = [strategy],
                            cash = 1000,
                            only_day=false)
start(trader)

Trader

Main task:    Task (failed) @0x00007fa65df0f3d0
Trading task: Task (done) @0x00007fa65df0f0f0
Data task:    Task (done) @0x00007fa65df0f260

Portfolio -- positions: 0, cash: 1000.0, tot: 1000.0

Current positions:
┌────────┬──────────┬───────┐
│ Ticker │ Quantity │ Value │
└────────┴──────────┴───────┘

Strategies:

Trades:
┌──────┬────────┬──────┬──────────┬───────────┬───────────┐
│ Time │ Ticker │ Side │ Quantity │ Avg Price │ Tot Price │
└──────┴────────┴──────┴──────────┴───────────┴───────────┘

Ledger
Components:
  0-element Component{Filled}
  184-element Component{PortfolioSnapshot}
  0-element Component{Position}
  0-element Component{Purchase}
  0-element Component{SMA{200, Close}}
  0-element Component{SMA{50, Close}}
  0-element Component{Sale}
  1-element Component{Strategy}
  184-element Component{TimeStamp}
  1-element Component{Trading.Cash}
  1-element Component{Trading.Clock}
  0-element Component{Trading.Order}
  1-element Component{Trading.PurchasePower}
Total entities: 188

To perform further analysis we can transform the trader data into a standard TimeArray as:

ta = TimeArray(trader)

1492×15 TimeArray{Float64, 2, DateTime, Matrix{Float64}} 2015-12-01T06:00:00 to 2019-12-31T06:00:00
│                     │ portfolio_value │ MSFT_SMA{200, Close} │ MSFT_Low │
├─────────────────────┼─────────────────┼──────────────────────┼──────────┤
│ 2015-12-01T06:00:00 │ 1000.0          │ NaN                  │ 54.3     │
│ 2015-12-02T06:00:00 │ 1000.0          │ NaN                  │ 55.06    │
│ 2015-12-03T06:00:00 │ 1000.0          │ NaN                  │ 53.93    │
│ 2015-12-04T06:00:00 │ 1000.0          │ NaN                  │ 54.1     │
│ 2015-12-05T06:00:00 │ 1000.0          │ NaN                  │ 54.4967  │
│ 2015-12-06T06:00:00 │ 1000.0          │ NaN                  │ 54.8933  │
│ 2015-12-07T06:00:00 │ 1000.0          │ NaN                  │ 55.29    │
│ 2015-12-08T06:00:00 │ 1000.0          │ NaN                  │ 54.99    │
│ 2015-12-09T06:00:00 │ 1000.0          │ NaN                  │ 54.51    │
   ⋮
│ 2019-12-24T06:00:00 │ NaN             │ 140.8999             │ 157.115  │
│ 2019-12-25T06:00:00 │ NaN             │ 141.0311             │ 157.2575 │
│ 2019-12-26T06:00:00 │ NaN             │ 141.1634             │ 157.4    │
│ 2019-12-27T06:00:00 │ NaN             │ 141.2952             │ 158.22   │
│ 2019-12-28T06:00:00 │ NaN             │ 141.4272             │ 157.7233 │
│ 2019-12-29T06:00:00 │ NaN             │ 141.56               │ 157.2267 │
│ 2019-12-30T06:00:00 │ NaN             │ 141.6864             │ 156.73   │
│ 2019-12-31T06:00:00 │ NaN             │ 141.8126             │ 156.45   │

│                     │ MSFT_Volume │ MSFT_Open │ MSFT_High │ MSFT_Close │
├─────────────────────┼─────────────┼───────────┼───────────┼────────────┤
│ 2015-12-01T06:00:00 │ 3.9952779e7 │ 54.41     │ 55.23     │ 55.22      │
│ 2015-12-02T06:00:00 │ 4.7274879e7 │ 55.32     │ 55.96     │ 55.21      │
│ 2015-12-03T06:00:00 │ 3.8627835e7 │ 55.49     │ 55.77     │ 54.2       │
│ 2015-12-04T06:00:00 │ 4.3963662e7 │ 54.12     │ 56.23     │ 55.91      │
│ 2015-12-05T06:00:00 │ 3.9545696e7 │ 54.6767   │ 56.1433   │ 55.8767    │
│ 2015-12-06T06:00:00 │ 3.5127731e7 │ 55.2333   │ 56.0567   │ 55.8433    │
│ 2015-12-07T06:00:00 │ 3.0709765e7 │ 55.79     │ 55.97     │ 55.81      │
│ 2015-12-08T06:00:00 │ 3.2878026e7 │ 55.47     │ 56.1      │ 55.79      │
│ 2015-12-09T06:00:00 │ 3.6353909e7 │ 55.37     │ 55.87     │ 54.98      │
   ⋮
│ 2019-12-24T06:00:00 │ 8.989184e6  │ 157.48    │ 157.71    │ 157.38     │
│ 2019-12-25T06:00:00 │ 1.1758237e7 │ 157.52    │ 158.22    │ 158.025    │
│ 2019-12-26T06:00:00 │ 1.452729e7  │ 157.56    │ 158.73    │ 158.67     │
│ 2019-12-27T06:00:00 │ 1.8414627e7 │ 159.45    │ 159.55    │ 158.96     │
│ 2019-12-28T06:00:00 │ 1.7728742e7 │ 159.2955  │ 159.3733  │ 158.5033   │
│ 2019-12-29T06:00:00 │ 1.7042856e7 │ 159.141   │ 159.1967  │ 158.0467   │
│ 2019-12-30T06:00:00 │ 1.6356971e7 │ 158.9865  │ 159.02    │ 157.59     │
│ 2019-12-31T06:00:00 │ 1.8393419e7 │ 156.77    │ 157.77    │ 157.7      │

│                     │ MSFT_SMA{50, Close} │ AAPL_SMA{200, Close} │ AAPL_Low │
├─────────────────────┼─────────────────────┼──────────────────────┼──────────┤
│ 2015-12-01T06:00:00 │ NaN                 │ NaN                  │ 116.86   │
│ 2015-12-02T06:00:00 │ NaN                 │ NaN                  │ 116.08   │
│ 2015-12-03T06:00:00 │ NaN                 │ NaN                  │ 114.22   │
│ 2015-12-04T06:00:00 │ NaN                 │ NaN                  │ 115.11   │
│ 2015-12-05T06:00:00 │ NaN                 │ NaN                  │ 116.01   │
│ 2015-12-06T06:00:00 │ NaN                 │ NaN                  │ 116.91   │
│ 2015-12-07T06:00:00 │ NaN                 │ NaN                  │ 117.81   │
│ 2015-12-08T06:00:00 │ NaN                 │ NaN                  │ 116.86   │
│ 2015-12-09T06:00:00 │ NaN                 │ NaN                  │ 115.08   │
   ⋮
│ 2019-12-24T06:00:00 │ 151.0492            │ 227.4527             │ 282.9197 │
│ 2019-12-25T06:00:00 │ 151.3205            │ 227.9334             │ 283.8098 │
│ 2019-12-26T06:00:00 │ 151.6127            │ 228.4241             │ 284.7    │
│ 2019-12-27T06:00:00 │ 151.9067            │ 228.9102             │ 288.12   │
│ 2019-12-28T06:00:00 │ 152.1576            │ 229.388              │ 287.1533 │
│ 2019-12-29T06:00:00 │ 152.3983            │ 229.8718             │ 286.1867 │
│ 2019-12-30T06:00:00 │ 152.6289            │ 230.3586             │ 285.22   │
│ 2019-12-31T06:00:00 │ 152.8607            │ 230.8632             │ 289.52   │

│                     │ AAPL_Volume │ AAPL_Open │ AAPL_High │ AAPL_Close │
├─────────────────────┼─────────────┼───────────┼───────────┼────────────┤
│ 2015-12-01T06:00:00 │ 3.4852374e7 │ 118.75    │ 118.81    │ 117.34     │
│ 2015-12-02T06:00:00 │ 3.3385643e7 │ 117.05    │ 118.11    │ 116.28     │
│ 2015-12-03T06:00:00 │ 4.1560785e7 │ 116.55    │ 116.79    │ 115.2      │
│ 2015-12-04T06:00:00 │ 5.7776977e7 │ 115.29    │ 119.25    │ 119.03     │
│ 2015-12-05T06:00:00 │ 4.9211569e7 │ 116.52    │ 119.4533  │ 118.78     │
│ 2015-12-06T06:00:00 │ 4.0646162e7 │ 117.75    │ 119.6567  │ 118.53     │
│ 2015-12-07T06:00:00 │ 3.2080754e7 │ 118.98    │ 119.86    │ 118.28     │
│ 2015-12-08T06:00:00 │ 3.430945e7  │ 117.52    │ 118.6     │ 118.23     │
│ 2015-12-09T06:00:00 │ 4.5852027e7 │ 117.64    │ 117.69    │ 115.62     │
   ⋮
│ 2019-12-24T06:00:00 │ 1.2119894e7 │ 284.69    │ 284.89    │ 284.27     │
│ 2019-12-25T06:00:00 │ 1.7727679e7 │ 284.755   │ 287.435   │ 287.09     │
│ 2019-12-26T06:00:00 │ 2.3335464e7 │ 284.82    │ 289.98    │ 289.91     │
│ 2019-12-27T06:00:00 │ 3.6593433e7 │ 291.12    │ 293.97    │ 289.8      │
│ 2019-12-28T06:00:00 │ 3.654568e7  │ 290.5667  │ 293.5433  │ 290.3733   │
│ 2019-12-29T06:00:00 │ 3.6497926e7 │ 290.0133  │ 293.1167  │ 290.9467   │
│ 2019-12-30T06:00:00 │ 3.6450173e7 │ 289.46    │ 292.69    │ 291.52     │
│ 2019-12-31T06:00:00 │ 2.558906e7  │ 289.93    │ 293.68    │ 293.65     │

by using Plots we can then plot certain columns in the TimeArray, e.g. the portfolio value:

using Plots
plot(ta[:portfolio_value])

We can see that this strategy is not particularly succesful.

See Slow Fast Strategy for a full runnable version of this strategy.

References

Trading.Strategy — Type

A Stage with a set of Systems that execute a strategy.

Trading.new_entities — Function

new_entities(ledger, strategy)

Returns a NewEntitiesIterator which iterates through the entities that have components that are requested by strategy, and were not yet seen. I.e. each entity in those components will be looped over once and only once when iteratively calling new_entities.

Trading.NewEntitiesIterator — Type

NewEntitiesIterator

Iterates through all the entities in the requested components of a Strategy that were not yet seen.