Trading

Trader

Trading.Trader — Type

Trader(broker::AbstractBroker; strategies::Vector{Pair{String, Vector{Strategy}}}, start=current_time())

Holds all data and tasks related to trading.

Trading.BackTester — Function

BackTester(broker)

Brokers

Trading.AbstractBroker — Type

AbstractBroker

Interface for external brokers.

Trading.AlpacaBroker — Type

AlpacaBroker

Broker to communicate with Alpaca. Can be constructed with your key_id and secret_key (see connect-to-alpaca-api).

Trading.HistoricalBroker — Type

HistoricalBroker

Stores and provides data from historical datasets. Data can be streamed fashion by assigning a Clock to the clock constructor kwarg, which will be used to determine the next bar to stream when calling receive on this broker.

Trading.MockBroker — Type

MockBroker

Mimics all function of a normal broker but with random data.