Data

Historical Acquisition

The current suggested method of historical data acquisition is through the AlpacaBroker. There are three types of data that can be retrieved: bars, quotes and trades.

Trading.bars — Function

bars(broker, ticker, start, stop; timeframe, kwargs...)

Retrieve the bar data for ticker from start to stop and with an interval of timeframe. When using AlpacaBroker see the Bar Object documentation for further reference.

Example

broker = AlpacaBroker(<key_id>, <secret_key>)

bars(broker, "AAPL", DateTime("2022-01-01T00:00:00"), DateTime("2023-01-01T00:00:00")l timeframe = Minute(1))

The above will retrieve 2022 bar data "AAPL" on a Minute resolution.

Trading.quotes — Function

quotes(broker, ticker, start, stop)

Returns the quotes made for ticker between start and stop. When using AlpacaBroker see the Quote Object documentation for further reference.

Example

broker = AlpacaBroker(<key_id>, <secret_key>)

quotes(broker, "AAPL", DateTime("2022-01-01T14:30:00"), DateTime("2022-01-01T14:31:00"))

Trading.trades — Function

trades(broker, ticker, start, stop)

Returns the trades made for ticker between start and stop. When using AlpacaBroker see the Trade Object documentation for further reference.

Example

broker = AlpacaBroker(<key_id>, <secret_key>)

trades(broker, "AAPL", DateTime("2022-01-01T14:30:00"), DateTime("2022-01-01T14:31:00"))

Each of these will return a TimeArray, which can be used e.g. with MarketTechnicals for specialized analysis. An AbstractBroker has an internal cache that will retain previously requested data.

Data Stream

There are realtime data streams for bars and portfolio/order updates. They follow the same semantics as the standard HTTP.WebSockets.WebSocket.

Trading.BarStream — Type

BarStream

Supplies a stream of bars from a broker. Can be created by calling bar_stream on an AbstractBroker. See receive and register! for more information.

Trading.bar_stream — Function

bar_stream(f::Function, broker)

Open a bar stream, calls function f with a BarStream object. Call receive on the BarStream to get new bars streamed in, and register! to register tickers for which to receive bar updates for.

HTTP.WebSockets.receive — Method

receive(barstream)

Blocking function which will return new bars as soon as they are available.

Trading.register! — Method

register!(barstream, ticker)

Register a ticker to the BarStream so that receive will also return updates with new bars for ticker.

Trading.OrderStream — Type

OrderStream

Interface to support executing trades and retrieving account updates.

Trading.order_stream — Function

order_stream(f::Function, broker::AbstractBroker)

Creates an OrderStream to stream order data. Uses the same semantics as a standard HTTP.WebSocket.

Example

broker = AlpacaBroker(<key_id>, <secret_key>)

order_stream(broker) do stream
    order = receive(stream)
end

Ticker Ledgers

Trading.TickerLedger — Type

TickerLedger

A TickerLedger holds the data for a given ticker as it arrives. Currently this is bar data in the form of Open, High, Low, Close and Volume, produced by a BarStream. If certain derived Indicator data is requested, it also holds this as it is produced by the different systems.

Bars

Trading.Open — Type

Open

The opening price of a given bar.

Trading.High — Type

High

The highest price of a given bar.

Trading.Low — Type

Low

The lowest price of a given bar.

Trading.Close — Type

Close

The closing price of a given bar.

Trading.Volume — Type

Volume

The traded volume of a given bar.

Indicators

Indicator data is generated from incoming bar data by the indicator systems as requested by the Strategy systems. Most of them have two type parameters designating the window or horizon of the indicator. For example the SMA{20, Close} closing price simple moving average indicator shows at a given timestamp the average of the closing prices of the 20 previous bars.

Accessing this data can be done through a Trader, e.g.:

trader = Trader(broker, tickers=["MSFT"])

trader["MSFT"][Trading.SMA{20, Trading.Close}]

provided that it was generated. See Strategies for more information.

Data

Trading.SMA — Type

SMA{horizon, T}

The simple moving average of a value over a timeframe of horizon.

Trading.MovingStdDev — Type

MovingStdDev{horizon, T}

The moving standard deviation of a value over timeframe of horizon.

Trading.EMA — Type

EMA{horizon, T}

The exponential moving average of a value over timeframe of horizon.

Trading.RSI — Type

RSI{horizon, T}

The relative strength index of a value over timeframe of horizon.

Trading.Bollinger — Type

Bollinger{horizon, T}

The up and down Bollinger bands for a value, over a timeframe of horizon.

Trading.Sharpe — Type

Sharpe{horizon, T}

The sharpe ratio of a value over a timeframe horizon.

Systems (@id indicator_systems)

Trading.SMACalculator — Type

SMACalculator

Calculates the SMA of data.

Trading.MovingStdDevCalculator — Type

MovingStdDevCalculator

Calculates the MovingStdDev of data.

Trading.EMACalculator — Type

EMACalculator

Calculates the EMA of data.

Trading.RSICalculator — Type

RSICalculator

Calculates the RSI of values.

Trading.BollingerCalculator — Type

BollingerCalculator

Calculates the Bollinger bands for data. The width parameter can be tuned, by default it is 2.0.

Trading.SharpeCalculator — Type

SharpeCalculator

Calculates the bare Sharpe ratio of data.

General Data

Trading.SingleValIndicator — Type

SingleValIndicator

A Component with a single value (usually v) that can be used by Indicator Systems to calculate various Indicators. If the single value is not stored in the v field of the Component, overload Trading.value.

Trading.LogVal — Type

LogVal

The logarithm of a value.

Trading.Difference — Type

Difference

The lag 1 difference.

Trading.RelativeDifference — Type

RelativeDifference

The lag 1 relative difference.

Trading.LogValCalculator — Type

LogValCalculator

Calculates the logarithms of data.

Trading.DifferenceCalculator — Type

DifferenceCalculator

Computes the lag 1 differences of data values.

Trading.RelativeDifferenceCalculator — Type

RelativeDifferenceCalculator

Calculates the lag 1 RelativeDifference of data.

Trading.value — Function

value(b::SingleValIndicator)

Returns the number that is stored in the SingleValIndicator. This is by default the v field.